Fixed PHPDocumentation
Our application was expecting a boolean to be returned by the auth() function. However, this seems to have been changed overtime. How is determined whether an authentication is successful? It seems the only way is to use 'checkToken()' after authentication. 

It would be better to have 'setToken()' return a boolean, however this might break other stuff that depends on the $this being returned now...